Improved cube check function
The program now returns 653 valid cube combinations. If we assume that
each of those combinations may be reversed so that cube1's layout may be
switched with cube2's layout to count as a new combination, that gives
us 1306 valid combinations. The correct answer is 1217, so we are still
off.